modufy codeblock to stringify children
